Universal_Profile_Codes_Prod3/scripts/universal_profile/recommendation_model/requirements.txt
PyYAML==5.3.1
appdirs==1.4.4
attrs==21.2.0
bcrypt==3.1.7
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.2.1
decorator==5.0.9
defusedxml==0.7.1
fsspec==0.7.4
gcsfs==0.6.2
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.28.0
google-cloud-core==1.4.4
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleads==28.0.0
googleapis-common-protos==1.52.0
httplib2==0.19.1
idna==2.10
isodate==0.6.0
joblib==0.14.1
lightfm==1.15
lxml==4.0.0
numpy==1.18.5
oauthlib==3.1.0
pandas==0.25.3
paramiko==2.7.2
pickle-mixin==1.0.2
protobuf==3.17.3
pyasn1-modules==0.2.8
pycparser==2.20
pynacl==1.4.0
pyparsing==2.4.7
pysftp==0.2.9
python-dateutil==2.8.1
pytz==2021.1
regex==2021.7.6
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
scikit-learn==0.22
scipy==1.4.1
setuptools==49.3.0
six==1.16.0
tqdm==4.61.2
uritemplate==3.0.1
urllib3==1.26.6
xgboost==0.90
xmltodict==0.12.0
zeep==3.4.0